+++ crates/ruma-client-api/src/sync/sync_events/v5.rs
@@ -20,8 +20,6 @@ use ruma_common::{
 use ruma_events::{AnyStrippedStateEvent, AnySyncStateEvent, AnySyncTimelineEvent, StateEventType};
 use serde::{Deserialize, Serialize};
 
-#[cfg(feature = "unstable-msc3575")]
-use super::v4;
 use super::UnreadNotificationsCount;
 
 const METADATA: Metadata = metadata! {
@@ -724,108 +722,6 @@ pub mod response {
         pub fn is_empty(&self) -> bool {
             self.rooms.is_empty()
         }
-    }
-}
-
-#[cfg(feature = "unstable-msc3575")]
-impl From<v4::Response> for Response {
-    fn from(value: v4::Response) -> Self {
-        Self {
-            pos: value.pos,
-            txn_id: value.txn_id,
-            lists: value.lists.into_iter().map(|(room_id, list)| (room_id, list.into())).collect(),
-            rooms: value.rooms.into_iter().map(|(room_id, room)| (room_id, room.into())).collect(),
-            extensions: value.extensions.into(),
-        }
-    }
-}
-
-#[cfg(feature = "unstable-msc3575")]
-impl From<v4::SyncList> for response::List {
-    fn from(value: v4::SyncList) -> Self {
-        Self { count: value.count }
-    }
-}
-
-#[cfg(feature = "unstable-msc3575")]
-impl From<v4::SlidingSyncRoom> for response::Room {
-    fn from(value: v4::SlidingSyncRoom) -> Self {
-        Self {
-            name: value.name,
-            avatar: value.avatar,
-            initial: value.initial,
-            is_dm: value.is_dm,
-            invite_state: value.invite_state,
-            unread_notifications: value.unread_notifications,
-            timeline: value.timeline,
-            required_state: value.required_state,
-            prev_batch: value.prev_batch,
-            limited: value.limited,
-            joined_count: value.joined_count,
-            invited_count: value.invited_count,
-            num_live: value.num_live,
-            bump_stamp: value.timestamp.map(|t| t.0),
-            heroes: value.heroes.map(|heroes| heroes.into_iter().map(Into::into).collect()),
-        }
-    }
-}
-
-#[cfg(feature = "unstable-msc3575")]
-impl From<v4::SlidingSyncRoomHero> for response::Hero {
-    fn from(value: v4::SlidingSyncRoomHero) -> Self {
-        Self { user_id: value.user_id, name: value.name, avatar: value.avatar }
-    }
-}
-
-#[cfg(feature = "unstable-msc3575")]
-impl From<v4::Extensions> for response::Extensions {
-    fn from(value: v4::Extensions) -> Self {
-        Self {
-            to_device: value.to_device.map(Into::into),
-            e2ee: value.e2ee.into(),
-            account_data: value.account_data.into(),
-            receipts: value.receipts.into(),
-            typing: value.typing.into(),
-        }
-    }
-}
-
-#[cfg(feature = "unstable-msc3575")]
-impl From<v4::ToDevice> for response::ToDevice {
-    fn from(value: v4::ToDevice) -> Self {
-        Self { next_batch: value.next_batch, events: value.events }
-    }
-}
-
-#[cfg(feature = "unstable-msc3575")]
-impl From<v4::E2EE> for response::E2EE {
-    fn from(value: v4::E2EE) -> Self {
-        Self {
-            device_lists: value.device_lists,
-            device_one_time_keys_count: value.device_one_time_keys_count,
-            device_unused_fallback_key_types: value.device_unused_fallback_key_types,
-        }
-    }
-}
-
-#[cfg(feature = "unstable-msc3575")]
-impl From<v4::AccountData> for response::AccountData {
-    fn from(value: v4::AccountData) -> Self {
-        Self { global: value.global, rooms: value.rooms }
-    }
-}
-
-#[cfg(feature = "unstable-msc3575")]
-impl From<v4::Receipts> for response::Receipts {
-    fn from(value: v4::Receipts) -> Self {
-        Self { rooms: value.rooms }
-    }
-}
-
-#[cfg(feature = "unstable-msc3575")]
-impl From<v4::Typing> for response::Typing {
-    fn from(value: v4::Typing) -> Self {
-        Self { rooms: value.rooms }
     }
 }