+++ crates/ruma-client-api/CHANGELOG.md
@@ -15,6 +15,8 @@ Improvements:
 - Add the `guest_access_token` field to `account::register::v3::Request` to
   allow a guest user to upgrade to a regular account.
 - Add the endpoints for peeking: `get_current_state` and `listen_to_new_events`.
+- Remove the unstable support for MSC3575 as it was closed. MSC4186 should be
+  used instead.
 
 # 0.20.1